backend/requirements.txt
pyautogen==0.2.38
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.9
httpx==0.27.0
